Float UMINUS in assembler

+++ tests/simple.c
@@ -1,10 +1,10 @@
 int main()
 {
-    int a = 2;
-    int b = 5;
-    int c;
+    float a = 2.0;
+    float b = 5.0;
+    float c;
 
-    c = (-a + b) * 2;
+    c = (-a + b) * 2.0;
     print("Hello world!");
     print(c);
 }
